fix: handle view more button
